[149] Delete old chessmove lambda
