--- lambda/chessmove/.gitignore
@@ -1,2 +1,0 @@
-lambda.zip
-output.json
