+++ lambda/chessmove/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
-name = "chessmove-lambda"
+name = "chessmove"
 version = "0.0.0"
-authors = ["Thomas Ball <tomoliverball@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -14,6 +13,13 @@ tokio = "1.0"
 serde_derive = "1.0.115"
 serde = "1.0.115"
 serde_json = "1.0.59"
-simple-error = "0.2.2"
 simple_logger = "1.16.0"
+itertools = "0.8"
 log = "0.4.14"
+bytes = "1.0"
+rand = "0.5.0"
+async-trait = "0.1.52"
+reqwest = { version = "0.11.8", default_features = false, features = ["rustls-tls", "blocking", "json"] }
+rusoto_core = { version = "0.47.0", default_features = false, features = ["rustls"] }
+rusoto_lambda = { version = "0.47.0", default_features = false, features = ["rustls"] }
+rusoto_dynamodb = { version = "0.47.0", default_features = false, features = ["rustls"] }
