+++ lambda/chessmove/src/endings.rs
