+++ lambda/chessmove/src/timing.rs
